feat: restructure app layout and add weather components
- Removed AppHeader from App.vue and integrated it into DefaultLayout.vue.
- Added new weather-related components: BasicWeatherInfo, DailyChart, PlanetaryInfo, WeatherForecast, and AppHeader.
- Updated routing to use DefaultLayout for the home view, enhancing the overall application structure.
