+++ src/App.vue
@@ -1,13 +1,9 @@
 <template>
-  <AppHeader />
   <RouterView />
 </template>
 
 <script>
-import AppHeader from './components/AppHeader.vue'
-
 export default {
-  name: 'App',
-  components: { AppHeader }
+  name: 'App'
 }
 </script>
